Reorganizing contrib subpackage
