+++ pyinterprod/interpro/contrib/common.py
@@ -1,0 +1,51 @@
+# -*- coding: utf-8 -*-
+
+from dataclasses import dataclass, field
+from datetime import datetime
+import xml.etree.ElementTree as ET
+from typing import List
+
+
+@dataclass
+class Method:
+    accession: str
+    sig_type: str
+    name: str = None
+    description: str = None
+    abstract: str = None
+    date: datetime = None
+
+
+@dataclass
+class Clan:
+    accession: str
+    name: str = None
+    description: str = None
+    members: list = field(default_factory=list)
+
+
+def parse_xml(filepath: str, sig_type: str) -> List[Method]:
+    """
+    Parse the interpro XML file provided by member databases
+
+    :param filepath: path the XML file
+    :param sig_type: signature type
+    :return:
+    """
+
+    with open(filepath, "rt", errors="replace") as fh:
+        tree = ET.parse(filepath)
+
+    root = tree.getroot()
+    namespace = "{http://www.ebi.ac.uk/schema/interpro}"
+    signatures = []
+    for sig in root.findall(f"{namespace}signature"):
+        attrib = sig.attrib
+        abstract = sig.find(f"{namespace}abstract").text.strip()
+        signatures.append(Method(accession=attrib["ac"],
+                                 sig_type=sig_type,
+                                 name=attrib["name"].replace(',', '_'),
+                                 description=attrib["desc"],
+                                 abstract=abstract))
+
+    return signatures
